src/backend/requirements.txt
fastapi
uvicorn
opencv-python
torch
ultralytics
google-generativeai
numpy
python-multipart
websockets
